Extract entity header visibility logic into hook, convert to TS
